Add Git Sources
Some sources need to handle Git operations
like cloning a repo, fetching latest changes, etc.

Since it's common for multiple types of sources (e.g. jjb, zuul.d)
it makes sense to create a Git-based source from which these
sources can inherit. Git source itself inherits from the generic
Source class.
