+++ requirements.txt
@@ -5,3 +5,4 @@ colorlog~=6.6.0
 urllib3~=1.26.9
 rfc3987~=1.3.8
 overrides~=6.1.0
+GitPython~=3.1.18
